Refactor search page and show no providers by default
